refactored UI - ready to be shown

+++ Udajovky/Data/Models/GpsPossition.swift
@@ -14,15 +14,15 @@ struct GpsPossition {
 //    let width: GpsWidth
 //    let length: GpsLength
     // TODO: isthis really needed ? lat/ long is defined on both - possitive and negative intervals, so it  is abvious
-    var lattitude: Int
-    let longitude: Int
+    var lattitude: Double
+    let longitude: Double
     // FIXME: double
 
     init(
         //        width: GpsWidth,
 //        length: GpsLength,
-        lattitude: Int,
-        longitude: Int
+        lattitude: Double,
+        longitude: Double
     ) {
 //        self.width = width
 //        self.length = length
